+++ tests/testthat/test-NormData.R
@@ -5,5 +5,9 @@ test_that("NormData", {
   normCount <- NormData(seuratlist)
 
   expect_type(normCount, "list")
+  expect_true(all(vapply(normCount, function(x) {
+    norms <- sqrt(MatrixGenerics::colSums(x ^ 2))
+    all(abs(norms - 1) < 1e-8)
+  }, logical(1))))
 
 })